requirements.txt
odfpy==1.4.1
Pillow==10.4.0
pynput==1.7.7
PyQt5==5.15.11
python-docx==1.1.2
requests==2.32.3
textract==1.5.0
tqdm>=4.65.0